mysite/requirements.txt
dataclasses==0.7
Django==2.2.7
django-rest-framework-mongoengine==3.4.0
djangorestframework==3.10.3
djongo==1.2.36
mongoengine==0.18.2
pkg-resources==0.0.0
pymongo==3.9.0
pytz==2019.3
six==1.13.0
sqlparse==0.2.4
